requirements.txt
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
pycparser==2.19
PyNaCl==1.3.0
six==1.13.0
